# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Video Topic Graph Platform API
  description: |
    Open-source platform for video topic graph generation.
    All endpoints require OIDC authentication via Bearer token.
  version: 1.0.0
  contact:
    name: Video Topic Graph Platform
    url: https://github.com/video-topic-graph/platform
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /topics/{id}:
    get:
      summary: Get topic details
      tags: [Topics]
      operationId: getTopic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Topic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update topic
      description: Update topic title, summary, or keywords
      tags: [Topics]
      operationId: updateTopic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTopicRequest'
      responses:
        '200':
          description: Topic updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /topics/{id}/snippets:
    get:
      summary: Get snippets for a topic
      tags: [Topics]
      operationId: getTopicSnippets
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of snippets
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnippetListResponse'
  /topics/merge:
    post:
      summary: Merge multiple topics
      description: |
        Merges two or more topics into a single topic.
        Creates a new topic and marks source topics as merged.
      tags: [Topics]
      operationId: mergeTopics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeTopicsRequest'
      responses:
        '200':
          description: Topics merged
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /topics/{id}/split:
    post:
      summary: Split a topic at a timestamp
      tags: [Topics]
      operationId: splitTopic
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SplitTopicRequest'
      responses:
        '200':
          description: Topic split into two
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TopicResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC Bearer token from Keycloak
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    UpdateTopicRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        summary:
          type: string
          maxLength: 5000
        keywords:
          type: array
          items:
            type: string
          maxItems: 20
        importance_score:
          type: number
          minimum: 0
          maximum: 1
    MergeTopicsRequest:
      type: object
      required: [topic_ids]
      properties:
        topic_ids:
          type: array
          items:
            type: string
            format: uuid
          minItems: 2
          maxItems: 10
        new_title:
          type: string
          description: Optional custom title for merged topic
    SplitTopicRequest:
      type: object
      required: [split_at_ts]
      properties:
        split_at_ts:
          type: number
          description: Timestamp to split at (seconds)
        new_titles:
          type: array
          items:
            type: string
          minItems: 2
          maxItems: 2
...
    TopicResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        video_id:
          type: string
          format: uuid
        level:
          type: integer
        start_ts:
          type: number
        end_ts:
          type: number
        title:
          type: string
        summary:
          type: string
        keywords:
          type: array
          items:
            type: string
        parent_ids:
          type: array
          items:
            type: string
            format: uuid
        child_ids:
          type: array
          items:
            type: string
            format: uuid
        importance_score:
          type: number
        cluster_id:
          type: string
          nullable: true
        snippets:
          type: array
          items:
            $ref: '#/components/schemas/Snippet'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SnippetListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Snippet'
...
    Snippet:
      type: object
      properties:
        id:
          type: string
          format: uuid
        topic_id:
          type: string
          format: uuid
        start_ts:
          type: number
        end_ts:
          type: number
        duration_s:
          type: number
        video_url:
          type: string
          format: uri
        thumbnail_url:
          type: string
          format: uri
          nullable: true
        caption_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
        request_id:
          type: string